+++ generate.sh
@@ -33,66 +33,62 @@ function main() {
   generate "sf/merger/v1/merger.proto"
   generate "sf/search/v1/search.proto"
   generate "sf/firehose/v1/firehose.proto"
-  generate "grpc/health/v1/health.proto"
 
   echo "generate.sh - `date` - `whoami`" > $ROOT/last_generate.txt
   echo "streamingfast/proto revision: `GIT_DIR=$PROTO/.git git rev-parse HEAD`" >> $ROOT/last_generate.txt
 }
 
+# usage:
+# - generate <protoPath>
+# - generate <protoBasePath/> [<file.proto> ...]
 function generate() {
-    protoc -I$PROTO $1 --go_out=plugins=grpc,paths=source_relative:.
+    base=""
+    if [[ "$#" -gt 1 ]]; then
+      base="$1"; shift
+    fi
+
+    for file in "$@"; do
+      protoc -I$PROTO \
+        --go_out=. --go_opt=paths=source_relative \
+        --go-grpc_out=. --go-grpc_opt=paths=source_relative,require_unimplemented_servers=false \
+         $base$file
+    done
 }
 
 function checks() {
-  result_1_4_0_and_later=`printf "" | protoc-gen-go --version 2>&1 | grep -Eo 'unknown argument'`
-  if [[ $result_1_4_0_and_later == "unknown argument" ]]; then
-     # We are using github.com/golang/protobuf/protoc-gen-go@v1.4.0+ it's the correct version we want here
-     return
-  fi
-
   # The old `protoc-gen-go` did not accept any flags. Just using `protoc-gen-go --version` in this
-  # version waits forever. So we pipe some wrong input to make it exit fast. This in the new version
+  # version waits forever. So we pipe some wrong input to make it exit fast. This, in the new version
   # which supports `--version` correctly print the version anyway and discard the standard input
   # so it's good with both version.
-  result_1_3_5_and_older=`printf "" | protoc-gen-go --version 2>&1 | grep -Eo v[0-9\.]+`
-  if [[ "$result_1_3_5_and_older" == "" ]]; then
+  result=`printf "" | protoc-gen-go --version 2>&1 | grep -Eo "v1\.(27)\.[0-9]+"`
+  if [[ "$result" == "" ]]; then
     echo "Your version of 'protoc-gen-go' (at `which protoc-gen-go`) is not recent enough."
     echo ""
-    echo "To fix your problem, perform those commands:"
+    echo "To fix your problem, perform this command (assumes you have Golang 1.17+):"
     echo ""
-    echo "  pushd /tmp"
-    echo "    go install github.com/golang/protobuf/protoc-gen-go@v1.5.2"
-    echo "  popd"
-    echo ""
-    echo "If everything is working as expetcted, the command:"
-    echo ""
-    echo "  protoc-gen-go --version"
-    echo ""
-    echo "Should print 'protoc-gen-go: unknown argument "--version" (this program should be run by protoc, not directly)'"
-    exit 1
-  fi
-
-  if [[ "$result_1_3_5_and_older" != "" ]]; then
-    echo "Your version of 'protoc-gen-go' is **too** recent!"
-    echo ""
-    echo "This repository requires a strict gRPC version not higher than v1.29.1 however"
-    echo "the newer protoc-gen-go versions generates code compatible with v1.32 at the minimum."
-    echo ""
-    echo "To keep the compatibility until the transitive dependency TiKV is updated (through streamingfast/kvdb)"
-    echo "you must ue the older package which is hosted at 'github.com/golang/protobuf/protoc-gen-go' (you most"
-    echo "probably have 'google.golang.org/protobuf/cmd/protoc-gen-go')."
-    echo ""
-    echo "To fix your problem, perform those commands:"
-    echo ""
-    echo "  pushd /tmp"
-    echo "    go install github.com/golang/protobuf/protoc-gen-go@v1.5.2"
-    echo "  popd"
+    echo "  go install google.golang.org/protobuf/cmd/protoc-gen-go@v1.27.1"
     echo ""
     echo "If everything is working as expected, the command:"
     echo ""
     echo "  protoc-gen-go --version"
     echo ""
-    echo "Should print 'protoc-gen-go: unknown argument "--version" (this program should be run by protoc, not directly)'"
+    echo "Should print 'protoc-gen-go v1.27.1' (or another more recent version)"
+    exit 1
+  fi
+
+  result=`printf "" | protoc-gen-go-grpc --version 2>&1 | grep -Eo "1\.2\.[0-9]+"`
+  if [[ "$result" == "" ]]; then
+    echo "Your version of 'protoc-gen-go-grpc' (at `which protoc-gen-go-grpc`) is not recent enough."
+    echo ""
+    echo "To fix your problem, perform this command (assumes you have Golang 1.17+):"
+    echo ""
+    echo "  go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.2.0"
+    echo ""
+    echo "If everything is working as expected, the command:"
+    echo ""
+    echo "  protoc-gen-go-grpc --version"
+    echo ""
+    echo "Should print 'protoc-gen-go-grpc 1.2.0' (or another more recent version)"
     exit 1
   fi
 }